+++ src/mapnik_vector_tile.cpp
@@ -1329,635 +1329,6 @@ Napi::Value VectorTile::painted(Napi::CallbackInfo const& info)
 }
 
 
-
-/*
-
-*/
-
-/*
-bool layer_to_geojson(protozero::pbf_reader const& layer,
-                      std::string & result,
-                      unsigned x,
-                      unsigned y,
-                      unsigned z)
-{
-    mapnik::vector_tile_impl::tile_datasource_pbf ds(layer, x, y, z);
-    mapnik::projection wgs84("+init=epsg:4326",true);
-    mapnik::projection merc("+init=epsg:3857",true);
-    mapnik::proj_transform prj_trans(merc,wgs84);
-    // This mega box ensures we capture all features, including those
-    // outside the tile extent. Geometries outside the tile extent are
-    // likely when the vtile was created by clipping to a buffered extent
-    mapnik::query q(mapnik::box2d<double>(std::numeric_limits<double>::lowest(),
-                                          std::numeric_limits<double>::lowest(),
-                                          std::numeric_limits<double>::max(),
-                                          std::numeric_limits<double>::max()));
-    mapnik::layer_descriptor ld = ds.get_descriptor();
-    for (auto const& item : ld.get_descriptors())
-    {
-        q.add_property_name(item.get_name());
-    }
-    mapnik::featureset_ptr fs = ds.features(q);
-    bool first = true;
-    if (fs && mapnik::is_valid(fs))
-    {
-        mapnik::feature_ptr feature;
-        while ((feature = fs->next()))
-        {
-            if (first)
-            {
-                first = false;
-            }
-            else
-            {
-                result += "\n,";
-            }
-            std::string feature_str;
-            mapnik::feature_impl feature_new(feature->context(),feature->id());
-            feature_new.set_data(feature->get_data());
-            unsigned int n_err = 0;
-            feature_new.set_geometry(mapnik::geometry::reproject_copy(feature->get_geometry(), prj_trans, n_err));
-            if (!mapnik::util::to_geojson(feature_str, feature_new))
-            {
-                // LCOV_EXCL_START
-                throw std::runtime_error("Failed to generate GeoJSON geometry");
-                // LCOV_EXCL_STOP
-            }
-            result += feature_str;
-        }
-    }
-    return !first;
-}
-*/
-/**
- * Syncronous version of {@link VectorTile}
- *
- * @memberof VectorTile
- * @instance
- * @name toGeoJSONSync
- * @param {string | number} [layer=__all__] Can be a zero-index integer representing
- * a layer or the string keywords `__array__` or `__all__` to get all layers in the form
- * of an array of GeoJSON `FeatureCollection`s or in the form of a single GeoJSON
- * `FeatureCollection` with all layers smooshed inside
- * @returns {string} stringified GeoJSON of all the features in this tile.
- * @example
- * var geojson = vectorTile.toGeoJSONSync('__all__');
- * geojson // stringified GeoJSON
- * JSON.parse(geojson); // GeoJSON object
- */
-Napi::Value VectorTile::toGeoJSONSync(Napi::CallbackInfo const& info)
-{
-    Napi::Env env = info.Env();
-    Napi::EscapableHandleScope scope(env);
-    return env.Undefined();
-}
-/*
-void write_geojson_array(std::string & result,
-                         VectorTile * v)
-{
-    protozero::pbf_reader tile_msg = v->get_tile()->get_reader();
-    result += "[";
-    bool first = true;
-    while (tile_msg.next(mapnik::vector_tile_impl::Tile_Encoding::LAYERS))
-    {
-        if (first)
-        {
-            first = false;
-        }
-        else
-        {
-            result += ",";
-        }
-        auto data_view = tile_msg.get_view();
-        protozero::pbf_reader layer_msg(data_view);
-        protozero::pbf_reader name_msg(data_view);
-        std::string layer_name;
-        if (name_msg.next(mapnik::vector_tile_impl::Layer_Encoding::NAME))
-        {
-            layer_name = name_msg.get_string();
-        }
-        result += "{\"type\":\"FeatureCollection\",";
-        result += "\"name\":\"" + layer_name + "\",\"features\":[";
-        std::string features;
-        bool hit = layer_to_geojson(layer_msg,
-                                    features,
-                                    v->get_tile()->x(),
-                                    v->get_tile()->y(),
-                                    v->get_tile()->z());
-        if (hit)
-        {
-            result += features;
-        }
-        result += "]}";
-    }
-    result += "]";
-}
-
-void write_geojson_all(std::string & result,
-                       VectorTile * v)
-{
-    protozero::pbf_reader tile_msg = v->get_tile()->get_reader();
-    result += "{\"type\":\"FeatureCollection\",\"features\":[";
-    bool first = true;
-    while (tile_msg.next(mapnik::vector_tile_impl::Tile_Encoding::LAYERS))
-    {
-        protozero::pbf_reader layer_msg(tile_msg.get_message());
-        std::string features;
-        bool hit = layer_to_geojson(layer_msg,
-                                    features,
-                                    v->get_tile()->x(),
-                                    v->get_tile()->y(),
-                                    v->get_tile()->z());
-        if (hit)
-        {
-            if (first)
-            {
-                first = false;
-            }
-            else
-            {
-                result += ",";
-            }
-            result += features;
-        }
-    }
-    result += "]}";
-}
-
-bool write_geojson_layer_index(std::string & result,
-                               std::size_t layer_idx,
-                               VectorTile * v)
-{
-    protozero::pbf_reader layer_msg;
-    if (v->get_tile()->layer_reader(layer_idx, layer_msg) &&
-        v->get_tile()->get_layers().size() > layer_idx)
-    {
-        std::string layer_name = v->get_tile()->get_layers()[layer_idx];
-        result += "{\"type\":\"FeatureCollection\",";
-        result += "\"name\":\"" + layer_name + "\",\"features\":[";
-        layer_to_geojson(layer_msg,
-                         result,
-                         v->get_tile()->x(),
-                         v->get_tile()->y(),
-                         v->get_tile()->z());
-        result += "]}";
-        return true;
-    }
-    // LCOV_EXCL_START
-    return false;
-    // LCOV_EXCL_STOP
-}
-
-bool write_geojson_layer_name(std::string & result,
-                              std::string const& name,
-                              VectorTile * v)
-{
-    protozero::pbf_reader layer_msg;
-    if (v->get_tile()->layer_reader(name, layer_msg))
-    {
-        result += "{\"type\":\"FeatureCollection\",";
-        result += "\"name\":\"" + name + "\",\"features\":[";
-        layer_to_geojson(layer_msg,
-                         result,
-                         v->get_tile()->x(),
-                         v->get_tile()->y(),
-                         v->get_tile()->z());
-        result += "]}";
-        return true;
-    }
-    return false;
-}
-*/
-/*
-Napi::Value VectorTile::_toGeoJSONSync(Napi::CallbackInfo const& info)
-{
-    Napi::EscapableHandleScope scope(env);
-    if (info.Length() < 1)
-    {
-        Napi::Error::New(env, "first argument must be either a layer name (string) or layer index (integer)").ThrowAsJavaScriptException();
-
-        return scope.Escape(env.Undefined());
-    }
-    Napi::Value layer_id = info[0];
-    if (! (layer_id.IsString() || layer_id.IsNumber()) )
-    {
-        Napi::TypeError::New(env, "'layer' argument must be either a layer name (string) or layer index (integer)").ThrowAsJavaScriptException();
-
-        return scope.Escape(env.Undefined());
-    }
-
-    VectorTile* v = info.Holder().Unwrap<VectorTile>();
-    std::string result;
-    try
-    {
-        if (layer_id.IsString())
-        {
-            std::string layer_name = TOSTR(layer_id);
-            if (layer_name == "__array__")
-            {
-                write_geojson_array(result, v);
-            }
-            else if (layer_name == "__all__")
-            {
-                write_geojson_all(result, v);
-            }
-            else
-            {
-                if (!write_geojson_layer_name(result, layer_name, v))
-                {
-                    std::string error_msg("Layer name '" + layer_name + "' not found");
-                    Napi::TypeError::New(env, error_msg.c_str()).ThrowAsJavaScriptException();
-
-                    return scope.Escape(env.Undefined());
-                }
-            }
-        }
-        else if (layer_id.IsNumber())
-        {
-            int layer_idx = layer_id.As<Napi::Number>().Int32Value();
-            if (layer_idx < 0)
-            {
-                Napi::TypeError::New(env, "A layer index can not be negative").ThrowAsJavaScriptException();
-
-                return scope.Escape(env.Undefined());
-            }
-            else if (layer_idx >= static_cast<int>(v->get_tile()->get_layers().size()))
-            {
-                Napi::TypeError::New(env, "Layer index exceeds the number of layers in the vector tile.").ThrowAsJavaScriptException();
-
-                return scope.Escape(env.Undefined());
-            }
-            if (!write_geojson_layer_index(result, layer_idx, v))
-            {
-                // LCOV_EXCL_START
-                Napi::TypeError::New(env, "Layer could not be retrieved (should have not reached here)").ThrowAsJavaScriptException();
-
-                return scope.Escape(env.Undefined());
-                // LCOV_EXCL_STOP
-            }
-        }
-    }
-    catch (std::exception const& ex)
-    {
-        // LCOV_EXCL_START
-        Napi::TypeError::New(env, ex.what()).ThrowAsJavaScriptException();
-
-        return scope.Escape(env.Undefined());
-        // LCOV_EXCL_STOP
-    }
-    return scope.Escape(Napi::String::New(env, result));
-}
-*/
-enum geojson_write_type : std::uint8_t
-{
-    geojson_write_all = 0,
-    geojson_write_array,
-    geojson_write_layer_name,
-    geojson_write_layer_index
-};
-/*
-struct to_geojson_baton
-{
-    uv_work_t request;
-    VectorTile* v;
-    bool error;
-    std::string result;
-    geojson_write_type type;
-    int layer_idx;
-    std::string layer_name;
-    Napi::FunctionReference cb;
-};
-*/
-/**
- * Get a [GeoJSON](http://geojson.org/) representation of this tile
- *
- * @memberof VectorTile
- * @instance
- * @name toGeoJSON
- * @param {string | number} [layer=__all__] Can be a zero-index integer representing
- * a layer or the string keywords `__array__` or `__all__` to get all layers in the form
- * of an array of GeoJSON `FeatureCollection`s or in the form of a single GeoJSON
- * `FeatureCollection` with all layers smooshed inside
- * @param {Function} callback - `function(err, geojson)`: a stringified
- * GeoJSON of all the features in this tile
- * @example
- * vectorTile.toGeoJSON('__all__',function(err, geojson) {
- *   if (err) throw err;
- *   console.log(geojson); // stringified GeoJSON
- *   console.log(JSON.parse(geojson)); // GeoJSON object
- * });
- */
-Napi::Value VectorTile::toGeoJSON(Napi::CallbackInfo const& info)
-{
-    Napi::Env env = info.Env();
-    //Napi::EscapableHandleScope scope(env);
-    return env.Undefined();
-/*
-    if ((info.Length() < 1) || !info[info.Length()-1]->IsFunction())
-    {
-        return _toGeoJSONSync(info);
-        return;
-    }
-    to_geojson_baton *closure = new to_geojson_baton();
-    closure->request.data = closure;
-    closure->v = info.Holder().Unwrap<VectorTile>();
-    closure->error = false;
-    closure->layer_idx = 0;
-    closure->type = geojson_write_all;
-
-    Napi::Value layer_id = info[0];
-    if (! (layer_id.IsString() || layer_id.IsNumber()) )
-    {
-        delete closure;
-        Napi::TypeError::New(env, "'layer' argument must be either a layer name (string) or layer index (integer)").ThrowAsJavaScriptException();
-        return env.Undefined();
-    }
-
-    if (layer_id.IsString())
-    {
-        std::string layer_name = TOSTR(layer_id);
-        if (layer_name == "__array__")
-        {
-            closure->type = geojson_write_array;
-        }
-        else if (layer_name == "__all__")
-        {
-            closure->type = geojson_write_all;
-        }
-        else
-        {
-            if (!closure->v->get_tile()->has_layer(layer_name))
-            {
-                delete closure;
-                std::string error_msg("The layer does not contain the name: " + layer_name);
-                Napi::TypeError::New(env, error_msg.c_str()).ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            closure->layer_name = layer_name;
-            closure->type = geojson_write_layer_name;
-        }
-    }
-    else if (layer_id.IsNumber())
-    {
-        closure->layer_idx = layer_id.As<Napi::Number>().Int32Value();
-        if (closure->layer_idx < 0)
-        {
-            delete closure;
-            Napi::TypeError::New(env, "A layer index can not be negative").ThrowAsJavaScriptException();
-            return env.Undefined();
-        }
-        else if (closure->layer_idx >= static_cast<int>(closure->v->get_tile()->get_layers().size()))
-        {
-            delete closure;
-            Napi::TypeError::New(env, "Layer index exceeds the number of layers in the vector tile.").ThrowAsJavaScriptException();
-            return env.Undefined();
-        }
-        closure->type = geojson_write_layer_index;
-    }
-
-    Napi::Value callback = info[info.Length()-1];
-    closure->cb.Reset(callback.As<Napi::Function>());
-    uv_queue_work(uv_default_loop(), &closure->request, to_geojson, (uv_after_work_cb)after_to_geojson);
-    closure->v->Ref();
-    return;
-*/
-}
-/*
-void VectorTile::to_geojson(uv_work_t* req)
-{
-    to_geojson_baton *closure = static_cast<to_geojson_baton *>(req->data);
-    try
-    {
-        switch (closure->type)
-        {
-            default:
-            case geojson_write_all:
-                write_geojson_all(closure->result, closure->v);
-                break;
-            case geojson_write_array:
-                write_geojson_array(closure->result, closure->v);
-                break;
-            case geojson_write_layer_name:
-                write_geojson_layer_name(closure->result, closure->layer_name, closure->v);
-                break;
-            case geojson_write_layer_index:
-                write_geojson_layer_index(closure->result, closure->layer_idx, closure->v);
-                break;
-        }
-    }
-    catch (std::exception const& ex)
-    {
-        // There are currently no known ways to trigger this exception in testing. If it was
-        // triggered this would likely be a bug in either mapnik or mapnik-vector-tile.
-        // LCOV_EXCL_START
-        closure->error = true;
-        closure->result = ex.what();
-        // LCOV_EXCL_STOP
-    }
-}
-
-void VectorTile::after_to_geojson(uv_work_t* req)
-{
-    Napi::HandleScope scope(env);
-    Napi::AsyncResource async_resource(__func__);
-    to_geojson_baton *closure = static_cast<to_geojson_baton *>(req->data);
-    if (closure->error)
-    {
-        // Because there are no known ways to trigger the exception path in to_geojson
-        // there is no easy way to test this path currently
-        // LCOV_EXCL_START
-        Napi::Value argv[1] = { Napi::Error::New(env, closure->result.c_str()) };
-        async_resource.runInAsyncScope(Napi::GetCurrentContext()->Global(), Napi::New(env, closure->cb), 1, argv);
-        // LCOV_EXCL_STOP
-    }
-    else
-    {
-        Napi::Value argv[2] = { env.Undefined(), Napi::String::New(env, closure->result) };
-        async_resource.runInAsyncScope(Napi::GetCurrentContext()->Global(), Napi::New(env, closure->cb), 2, argv);
-    }
-    closure->v->Unref();
-    closure->cb.Reset();
-    delete closure;
-}
-*/
-
-/**
- * Add features to this tile from a GeoJSON string. GeoJSON coordinates must be in the WGS84 longitude & latitude CRS
- * as specified in the [GeoJSON Specification](https://www.rfc-editor.org/rfc/rfc7946.txt).
- *
- * @memberof VectorTile
- * @instance
- * @name addGeoJSON
- * @param {string} geojson as a string
- * @param {string} name of the layer to be added
- * @param {Object} [options]
- * @param {number} [options.area_threshold=0.1] used to discard small polygons.
- * If a value is greater than `0` it will trigger polygons with an area smaller
- * than the value to be discarded. Measured in grid integers, not spherical mercator
- * coordinates.
- * @param {number} [options.simplify_distance=0.0] Simplification works to generalize
- * geometries before encoding into vector tiles.simplification distance The
- * `simplify_distance` value works in integer space over a 4096 pixel grid and uses
- * the [Douglas-Peucker algorithm](https://en.wikipedia.org/wiki/Ramer%E2%80%93Douglas%E2%80%93Peucker_algorithm).
- * @param {boolean} [options.strictly_simple=true] ensure all geometry is valid according to
- * OGC Simple definition
- * @param {boolean} [options.multi_polygon_union=false] union all multipolygons
- * @param {Object<mapnik.polygonFillType>} [options.fill_type=mapnik.polygonFillType.positive]
- * the fill type used in determining what are holes and what are outer rings. See the
- * [Clipper documentation](http://www.angusj.com/delphi/clipper/documentation/Docs/Units/ClipperLib/Types/PolyFillType.htm)
- * to learn more about fill types.
- * @param {boolean} [options.process_all_rings=false] if `true`, don't assume winding order and ring order of
- * polygons are correct according to the [`2.0` Mapbox Vector Tile specification](https://github.com/mapbox/vector-tile-spec)
- * @example
- * var geojson = { ... };
- * var vt = mapnik.VectorTile(0,0,0);
- * vt.addGeoJSON(JSON.stringify(geojson), 'layer-name', {});
- */
-Napi::Value VectorTile::addGeoJSON(Napi::CallbackInfo const& info)
-{
-    Napi::Env env = info.Env();
-    //Napi::EscapableHandleScope scope(env);
-    return env.Undefined();
-/*
-
-    VectorTile* d = info.Holder().Unwrap<VectorTile>();
-    if (info.Length() < 1 || !info[0].IsString())
-    {
-        Napi::Error::New(env, "first argument must be a GeoJSON string").ThrowAsJavaScriptException();
-        return env.Undefined();
-    }
-    if (info.Length() < 2 || !info[1].IsString())
-    {
-        Napi::Error::New(env, "second argument must be a layer name (string)").ThrowAsJavaScriptException();
-        return env.Undefined();
-    }
-    std::string geojson_string = TOSTR(info[0]);
-    std::string geojson_name = TOSTR(info[1]);
-
-    Napi::Object options = Napi::Object::New(env);
-    double area_threshold = 0.1;
-    double simplify_distance = 0.0;
-    bool strictly_simple = true;
-    bool multi_polygon_union = false;
-    mapnik::vector_tile_impl::polygon_fill_type fill_type = mapnik::vector_tile_impl::positive_fill;
-    bool process_all_rings = false;
-
-    if (info.Length() > 2)
-    {
-        // options object
-        if (!info[2].IsObject())
-        {
-            Napi::Error::New(env, "optional third argument must be an options object").ThrowAsJavaScriptException();
-            return env.Undefined();
-        }
-
-        options = info[2].ToObject(Napi::GetCurrentContext());
-        if ((options).Has(Napi::String::New(env, "area_threshold")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "area_threshold"));
-            if (!param_val.IsNumber())
-            {
-                Napi::Error::New(env, "option 'area_threshold' must be a number").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            area_threshold = param_val.As<Napi::Number>().Int32Value();
-            if (area_threshold < 0.0)
-            {
-                Napi::Error::New(env, "option 'area_threshold' can not be negative").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-        }
-        if ((options).Has(Napi::String::New(env, "strictly_simple")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "strictly_simple"));
-            if (!param_val->IsBoolean())
-            {
-                Napi::Error::New(env, "option 'strictly_simple' must be a boolean").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            strictly_simple = param_val.As<Napi::Boolean>().Value();
-        }
-        if ((options).Has(Napi::String::New(env, "multi_polygon_union")).FromMaybe(false))
-        {
-            Napi::Value mpu = (options).Get(Napi::String::New(env, "multi_polygon_union"));
-            if (!mpu->IsBoolean())
-            {
-                Napi::TypeError::New(env, "multi_polygon_union value must be a boolean").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            multi_polygon_union = mpu.As<Napi::Boolean>().Value();
-        }
-        if ((options).Has(Napi::String::New(env, "fill_type")).FromMaybe(false))
-        {
-            Napi::Value ft = (options).Get(Napi::String::New(env, "fill_type"));
-            if (!ft.IsNumber())
-            {
-                Napi::TypeError::New(env, "optional arg 'fill_type' must be a number").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            fill_type = static_cast<mapnik::vector_tile_impl::polygon_fill_type>(ft.As<Napi::Number>().Int32Value());
-            if (fill_type >= mapnik::vector_tile_impl::polygon_fill_type_max)
-            {
-                Napi::TypeError::New(env, "optional arg 'fill_type' out of possible range").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-        }
-        if ((options).Has(Napi::String::New(env, "simplify_distance")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "simplify_distance"));
-            if (!param_val.IsNumber())
-            {
-                Napi::TypeError::New(env, "option 'simplify_distance' must be an floating point number").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            simplify_distance = param_val.As<Napi::Number>().DoubleValue();
-            if (simplify_distance < 0.0)
-            {
-                Napi::TypeError::New(env, "option 'simplify_distance' must be a positive number").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-        }
-        if ((options).Has(Napi::String::New(env, "process_all_rings")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "process_all_rings"));
-            if (!param_val->IsBoolean())
-            {
-                Napi::TypeError::New(env, "option 'process_all_rings' must be a boolean").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            process_all_rings = param_val.As<Napi::Boolean>().Value();
-        }
-    }
-
-    try
-    {
-        // create map object
-        mapnik::Map map(d->tile_size(),d->tile_size(),"+init=epsg:3857");
-        mapnik::parameters p;
-        p["type"]="geojson";
-        p["inline"]=geojson_string;
-        mapnik::layer lyr(geojson_name,"+init=epsg:4326");
-        lyr.set_datasource(mapnik::datasource_cache::instance().create(p));
-        map.add_layer(lyr);
-
-        mapnik::vector_tile_impl::processor ren(map);
-        ren.set_area_threshold(area_threshold);
-        ren.set_strictly_simple(strictly_simple);
-        ren.set_simplify_distance(simplify_distance);
-        ren.set_multi_polygon_union(multi_polygon_union);
-        ren.set_fill_type(fill_type);
-        ren.set_process_all_rings(process_all_rings);
-        ren.update_tile(*d->get_tile());
-        return env.True();
-    }
-    catch (std::exception const& ex)
-    {
-        Napi::Error::New(env, ex.what()).ThrowAsJavaScriptException();
-        return env.Undefined();
-    }
-*/
-}
-
 /**
  * Add a {@link Image} as a tile layer (synchronous)
  *
@@ -2734,433 +2105,6 @@ void VectorTile::EIO_AfterAddData(uv_work_t* req)
     closure->d->Unref();
     closure->cb.Reset();
     closure->buffer.Reset();
-    delete closure;
-}
-*/
-
-/**
- * Get the data in this vector tile as a buffer (synchronous)
- *
- * @memberof VectorTile
- * @instance
- * @name getDataSync
- * @param {Object} [options]
- * @param {string} [options.compression=none] - can also be `gzip`
- * @param {boolean} [options.release=false] releases VT buffer
- * @param {int} [options.level=0] a number `0` (no compression) to `9` (best compression)
- * @param {string} options.strategy must be `FILTERED`, `HUFFMAN_ONLY`, `RLE`, `FIXED`, `DEFAULT`
- * @returns {Buffer} raw data
- * @example
- * var data = vt.getData({
- *   compression: 'gzip',
- *   level: 9,
- *   strategy: 'FILTERED'
- * });
- */
-Napi::Value VectorTile::getDataSync(Napi::CallbackInfo const& info)
-{
-    Napi::Env env = info.Env();
-    Napi::EscapableHandleScope scope(env);
-    return env.Undefined();
-    /*
-    VectorTile* d = info.Holder().Unwrap<VectorTile>();
-
-    bool compress = false;
-    bool release = false;
-    int level = Z_DEFAULT_COMPRESSION;
-    int strategy = Z_DEFAULT_STRATEGY;
-
-    Napi::Object options = Napi::Object::New(env);
-
-    if (info.Length() > 0)
-    {
-        if (!info[0].IsObject())
-        {
-            Napi::TypeError::New(env, "first arg must be a options object").ThrowAsJavaScriptException();
-
-            return scope.Escape(env.Undefined());
-        }
-
-        options = info[0].ToObject(Napi::GetCurrentContext());
-
-        if ((options).Has(Napi::String::New(env, "compression")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "compression"));
-            if (!param_val.IsString())
-            {
-                Napi::TypeError::New(env, "option 'compression' must be a string, either 'gzip', or 'none' (default)").ThrowAsJavaScriptException();
-
-                return scope.Escape(env.Undefined());
-            }
-            compress = std::string("gzip") == (TOSTR(param_val->ToString(Napi::GetCurrentContext())));
-        }
-        if ((options).Has(Napi::String::New(env, "release")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "release"));
-            if (!param_val->IsBoolean())
-            {
-                Napi::Error::New(env, "option 'release' must be a boolean").ThrowAsJavaScriptException();
-
-                return scope.Escape(env.Undefined());
-            }
-            release = param_val.As<Napi::Boolean>().Value();
-        }
-        if ((options).Has(Napi::String::New(env, "level")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "level"));
-            if (!param_val.IsNumber())
-            {
-                Napi::TypeError::New(env, "option 'level' must be an integer between 0 (no compression) and 9 (best compression) inclusive").ThrowAsJavaScriptException();
-
-                return scope.Escape(env.Undefined());
-            }
-            level = param_val.As<Napi::Number>().Int32Value();
-            if (level < 0 || level > 9)
-            {
-                Napi::TypeError::New(env, "option 'level' must be an integer between 0 (no compression) and 9 (best compression) inclusive").ThrowAsJavaScriptException();
-
-                return scope.Escape(env.Undefined());
-            }
-        }
-        if ((options).Has(Napi::String::New(env, "strategy")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "strategy"));
-            if (!param_val.IsString())
-            {
-                Napi::TypeError::New(env, "option 'strategy' must be one of the following strings: FILTERED, HUFFMAN_ONLY, RLE, FIXED, DEFAULT").ThrowAsJavaScriptException();
-
-                return scope.Escape(env.Undefined());
-            }
-            else if (std::string("FILTERED") == TOSTR(param_val->ToString(Napi::GetCurrentContext())))
-            {
-                strategy = Z_FILTERED;
-            }
-            else if (std::string("HUFFMAN_ONLY") == TOSTR(param_val->ToString(Napi::GetCurrentContext())))
-            {
-                strategy = Z_HUFFMAN_ONLY;
-            }
-            else if (std::string("RLE") == TOSTR(param_val->ToString(Napi::GetCurrentContext())))
-            {
-                strategy = Z_RLE;
-            }
-            else if (std::string("FIXED") == TOSTR(param_val->ToString(Napi::GetCurrentContext())))
-            {
-                strategy = Z_FIXED;
-            }
-            else if (std::string("DEFAULT") == TOSTR(param_val->ToString(Napi::GetCurrentContext())))
-            {
-                strategy = Z_DEFAULT_STRATEGY;
-            }
-            else
-            {
-                Napi::TypeError::New(env, "option 'strategy' must be one of the following strings: FILTERED, HUFFMAN_ONLY, RLE, FIXED, DEFAULT").ThrowAsJavaScriptException();
-
-                return scope.Escape(env.Undefined());
-            }
-        }
-    }
-
-    try
-    {
-        std::size_t raw_size = d->tile_->size();
-        if (raw_size <= 0)
-        {
-            return scope.Escape(Napi::Buffer<char>::New(env, 0));
-        }
-        else
-        {
-            if (raw_size >= node::Buffer::kMaxLength) {
-                // This is a valid test path, but I am excluding it from test coverage due to the
-                // requirement of loading a very large object in memory in order to test it.
-                // LCOV_EXCL_START
-                std::ostringstream s;
-                s << "Data is too large to convert to a node::Buffer ";
-                s << "(" << raw_size << " raw bytes >= node::Buffer::kMaxLength)";
-                Napi::TypeError::New(env, s.str().c_str()).ThrowAsJavaScriptException();
-
-                return scope.Escape(env.Undefined());
-                // LCOV_EXCL_STOP
-            }
-            if (!compress)
-            {
-                if (release)
-                {
-                    return scope.Escape(node_mapnik::NewBufferFrom(d->tile_->release_buffer()));
-                }
-                else
-                {
-                    return scope.Escape(Napi::Buffer::Copy(env, (char*)d->tile_->data(),raw_size));
-                }
-            }
-            else
-            {
-                std::unique_ptr<std::string> compressed = std::make_unique<std::string>();
-                mapnik::vector_tile_impl::zlib_compress(d->tile_->data(), raw_size, *compressed, true, level, strategy);
-                if (release)
-                {
-                    // To keep the same behaviour as a non compression release, we want to clear the VT buffer
-                    d->tile_->clear();
-                }
-                return scope.Escape(node_mapnik::NewBufferFrom(std::move(compressed)));
-            }
-        }
-    }
-    catch (std::exception const& ex)
-    {
-        // As all exception throwing paths are not easily testable or no way can be
-        // found to test with repeatability this exception path is not included
-        // in test coverage.
-        // LCOV_EXCL_START
-        Napi::TypeError::New(env, ex.what()).ThrowAsJavaScriptException();
-
-        return scope.Escape(env.Undefined());
-        // LCOV_EXCL_STOP
-    }
-    return scope.Escape(env.Undefined());
-    */
-}
-/*
-typedef struct
-{
-    uv_work_t request;
-    VectorTile* d;
-    bool error;
-    std::unique_ptr<std::string> data;
-    bool compress;
-    bool release;
-    int level;
-    int strategy;
-    std::string error_name;
-    Napi::FunctionReference cb;
-} vector_tile_get_data_baton_t;
-*/
-
-/**
- * Get the data in this vector tile as a buffer (asynchronous)
- * @memberof VectorTile
- * @instance
- * @name getData
- * @param {Object} [options]
- * @param {string} [options.compression=none] compression type can also be `gzip`
- * @param {boolean} [options.release=false] releases VT buffer
- * @param {int} [options.level=0] a number `0` (no compression) to `9` (best compression)
- * @param {string} options.strategy must be `FILTERED`, `HUFFMAN_ONLY`, `RLE`, `FIXED`, `DEFAULT`
- * @param {Function} callback
- * @example
- * vt.getData({
- *   compression: 'gzip',
- *   level: 9,
- *   strategy: 'FILTERED'
- * }, function(err, data) {
- *   if (err) throw err;
- *   console.log(data); // buffer
- * });
- */
-Napi::Value VectorTile::getData(Napi::CallbackInfo const& info)
-{
-    Napi::Env env = info.Env();
-    ///Napi::EscapableHandleScope scope(env);
-    return env.Undefined();
-    /*
-    if (info.Length() == 0 || !info[info.Length()-1]->IsFunction())
-    {
-        return _getDataSync(info);
-        return;
-    }
-
-    Napi::Value callback = info[info.Length()-1];
-    bool compress = false;
-    bool release = false;
-    int level = Z_DEFAULT_COMPRESSION;
-    int strategy = Z_DEFAULT_STRATEGY;
-
-    Napi::Object options = Napi::Object::New(env);
-
-    if (info.Length() > 1)
-    {
-        if (!info[0].IsObject())
-        {
-            Napi::TypeError::New(env, "first arg must be a options object").ThrowAsJavaScriptException();
-            return env.Undefined();
-        }
-
-        options = info[0].ToObject(Napi::GetCurrentContext());
-
-        if ((options).Has(Napi::String::New(env, "compression")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "compression"));
-            if (!param_val.IsString())
-            {
-                Napi::TypeError::New(env, "option 'compression' must be a string, either 'gzip', or 'none' (default)").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            compress = std::string("gzip") == (TOSTR(param_val->ToString(Napi::GetCurrentContext())));
-        }
-        if ((options).Has(Napi::String::New(env, "release")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "release"));
-            if (!param_val->IsBoolean())
-            {
-                Napi::TypeError::New(env, "option 'release' must be a boolean").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            release = param_val.As<Napi::Boolean>().Value();
-        }
-        if ((options).Has(Napi::String::New(env, "level")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "level"));
-            if (!param_val.IsNumber())
-            {
-                Napi::TypeError::New(env, "option 'level' must be an integer between 0 (no compression) and 9 (best compression) inclusive").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            level = param_val.As<Napi::Number>().Int32Value();
-            if (level < 0 || level > 9)
-            {
-                Napi::TypeError::New(env, "option 'level' must be an integer between 0 (no compression) and 9 (best compression) inclusive").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-        }
-        if ((options).Has(Napi::String::New(env, "strategy")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "strategy"));
-            if (!param_val.IsString())
-            {
-                Napi::TypeError::New(env, "option 'strategy' must be one of the following strings: FILTERED, HUFFMAN_ONLY, RLE, FIXED, DEFAULT").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            else if (std::string("FILTERED") == TOSTR(param_val->ToString(Napi::GetCurrentContext())))
-            {
-                strategy = Z_FILTERED;
-            }
-            else if (std::string("HUFFMAN_ONLY") == TOSTR(param_val->ToString(Napi::GetCurrentContext())))
-            {
-                strategy = Z_HUFFMAN_ONLY;
-            }
-            else if (std::string("RLE") == TOSTR(param_val->ToString(Napi::GetCurrentContext())))
-            {
-                strategy = Z_RLE;
-            }
-            else if (std::string("FIXED") == TOSTR(param_val->ToString(Napi::GetCurrentContext())))
-            {
-                strategy = Z_FIXED;
-            }
-            else if (std::string("DEFAULT") == TOSTR(param_val->ToString(Napi::GetCurrentContext())))
-            {
-                strategy = Z_DEFAULT_STRATEGY;
-            }
-            else
-            {
-                Napi::TypeError::New(env, "option 'strategy' must be one of the following strings: FILTERED, HUFFMAN_ONLY, RLE, FIXED, DEFAULT").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-        }
-    }
-
-    VectorTile* d = info.Holder().Unwrap<VectorTile>();
-    vector_tile_get_data_baton_t *closure = new vector_tile_get_data_baton_t();
-    closure->request.data = closure;
-    closure->d = d;
-    closure->compress = compress;
-    closure->release = release;
-    closure->data = std::make_unique<std::string>();
-    closure->level = level;
-    closure->strategy = strategy;
-    closure->error = false;
-    closure->cb.Reset(callback.As<Napi::Function>());
-    uv_queue_work(uv_default_loop(), &closure->request, get_data, (uv_after_work_cb)after_get_data);
-    d->Ref();
-    return;
-    */
-}
-/*
-void VectorTile::get_data(uv_work_t* req)
-{
-    vector_tile_get_data_baton_t *closure = static_cast<vector_tile_get_data_baton_t *>(req->data);
-    try
-    {
-        // compress if requested
-        if (closure->compress)
-        {
-            mapnik::vector_tile_impl::zlib_compress(closure->d->tile_->data(), closure->d->tile_->size(), *(closure->data), true, closure->level, closure->strategy);
-        }
-    }
-    catch (std::exception const& ex)
-    {
-        // As all exception throwing paths are not easily testable or no way can be
-        // found to test with repeatability this exception path is not included
-        // in test coverage.
-        // LCOV_EXCL_START
-        closure->error = true;
-        closure->error_name = ex.what();
-        // LCOV_EXCL_STOP
-    }
-}
-
-void VectorTile::after_get_data(uv_work_t* req)
-{
-    Napi::HandleScope scope(env);
-    Napi::AsyncResource async_resource(__func__);
-    vector_tile_get_data_baton_t *closure = static_cast<vector_tile_get_data_baton_t *>(req->data);
-    if (closure->error)
-    {
-        // As all exception throwing paths are not easily testable or no way can be
-        // found to test with repeatability this exception path is not included
-        // in test coverage.
-        // LCOV_EXCL_START
-        Napi::Value argv[1] = { Napi::Error::New(env, closure->error_name.c_str()) };
-        async_resource.runInAsyncScope(Napi::GetCurrentContext()->Global(), Napi::New(env, closure->cb), 1, argv);
-        // LCOV_EXCL_STOP
-    }
-    else if (!closure->data->empty())
-    {
-        if (closure->release)
-        {
-            // To keep the same behaviour as a non compression release, we want to clear the VT buffer
-            closure->d->tile_->clear();
-        }
-        Napi::Value argv[2] = { env.Undefined(),
-                                         node_mapnik::NewBufferFrom(std::move(closure->data)) };
-        async_resource.runInAsyncScope(Napi::GetCurrentContext()->Global(), Napi::New(env, closure->cb), 2, argv);
-    }
-    else
-    {
-        std::size_t raw_size = closure->d->tile_->size();
-        if (raw_size <= 0)
-        {
-            Napi::Value argv[2] = { env.Undefined(), Napi::Buffer<char>::New(env, 0) };
-            async_resource.runInAsyncScope(Napi::GetCurrentContext()->Global(), Napi::New(env, closure->cb), 2, argv);
-        }
-        else if (raw_size >= node::Buffer::kMaxLength)
-        {
-            // This is a valid test path, but I am excluding it from test coverage due to the
-            // requirement of loading a very large object in memory in order to test it.
-            // LCOV_EXCL_START
-            std::ostringstream s;
-            s << "Data is too large to convert to a node::Buffer ";
-            s << "(" << raw_size << " raw bytes >= node::Buffer::kMaxLength)";
-            Napi::Value argv[1] = { Napi::Error::New(env, s.str().c_str()) };
-            async_resource.runInAsyncScope(Napi::GetCurrentContext()->Global(), Napi::New(env, closure->cb), 1, argv);
-            // LCOV_EXCL_STOP
-        }
-        else
-        {
-            if (closure->release)
-            {
-                Napi::Value argv[2] = { env.Undefined(),
-                                                 node_mapnik::NewBufferFrom(closure->d->tile_->release_buffer()) };
-                async_resource.runInAsyncScope(Napi::GetCurrentContext()->Global(), Napi::New(env, closure->cb), 2, argv);
-            }
-            else
-            {
-                Napi::Value argv[2] = { env.Undefined(), Napi::Buffer::Copy(env, (char*)closure->d->tile_->data(),raw_size) };
-                async_resource.runInAsyncScope(Napi::GetCurrentContext()->Global(), Napi::New(env, closure->cb), 2, argv);
-            }
-        }
-    }
-
-    closure->d->Unref();
-    closure->cb.Reset();
     delete closure;
 }
 */
@@ -4954,7 +3898,6 @@ void VectorTile::set_tile_z(Napi::CallbackInfo const& info, const Napi::Value& v
 
 void VectorTile::set_tile_size(Napi::CallbackInfo const& info, const Napi::Value& value)
 {
-    //VectorTile* d = info.Holder().Unwrap<VectorTile>();
     Napi::Env env = info.Env();
     if (!value.IsNumber())
     {
@@ -4976,7 +3919,6 @@ void VectorTile::set_tile_size(Napi::CallbackInfo const& info, const Napi::Value
 void VectorTile::set_buffer_size(Napi::CallbackInfo const& info, const Napi::Value& value)
 {
     Napi::Env env = info.Env();
-    //VectorTile* d = info.Holder().Unwrap<VectorTile>();
     if (!value.IsNumber())
     {
         Napi::Error::New(env, "Must provide a number").ThrowAsJavaScriptException();
@@ -4985,203 +3927,11 @@ void VectorTile::set_buffer_size(Napi::CallbackInfo const& info, const Napi::Val
     else
     {
         int val = value.As<Napi::Number>().Int32Value();
-        if (static_cast<int>(tile_size()) + (2 * val) <= 0)
+        if (static_cast<int>(tile_->tile_size()) + (2 * val) <= 0)
         {
             Napi::Error::New(env, "too large of a negative buffer for tilesize").ThrowAsJavaScriptException();
             return;
         }
         tile_->buffer_size(val);
     }
-}
-/*
-typedef struct {
-    uv_work_t request;
-    const char *data;
-    size_t dataLength;
-    Napi::Object obj;
-    bool error;
-    std::string error_str;
-    Napi::Persistent<v8::Object> buffer;
-    Napi::FunctionReference cb;
-} vector_tile_info_baton_t;
-*/
-/**
- * Return an object containing information about a vector tile buffer. Useful for
- * debugging `.mvt` files with errors.
- *
- * @name info
- * @param {Buffer} buffer - vector tile buffer
- * @returns {Object} json object with information about the vector tile buffer
- * @static
- * @memberof VectorTile
- * @instance
- * @example
- * var buffer = fs.readFileSync('./path/to/tile.mvt');
- * var info = mapnik.VectorTile.info(buffer);
- * console.log(info);
- * // { layers:
- * //   [ { name: 'world',
- * //      features: 1,
- * //      point_features: 0,
- * //      linestring_features: 0,
- * //      polygon_features: 1,
- * //      unknown_features: 0,
- * //      raster_features: 0,
- * //      version: 2 },
- * //    { name: 'world2',
- * //      features: 1,
- * //      point_features: 0,
- * //      linestring_features: 0,
- * //      polygon_features: 1,
- * //      unknown_features: 0,
- * //      raster_features: 0,
- * //      version: 2 } ],
- * //    errors: false }
- */
-Napi::Value VectorTile::info(Napi::CallbackInfo const& info)
-{
-    Napi::Env env = info.Env();
-    Napi::EscapableHandleScope scope(env);
-    return env.Undefined();
-
-    /*
-    if (info.Length() < 1 || !info[0].IsObject())
-    {
-        Napi::TypeError::New(env, "must provide a buffer argument").ThrowAsJavaScriptException();
-        return env.Undefined();
-    }
-
-    Napi::Object obj = info[0].ToObject(Napi::GetCurrentContext());
-    if (!obj.IsBuffer())
-    {
-        Napi::TypeError::New(env, "first argument is invalid, must be a Buffer").ThrowAsJavaScriptException();
-        return env.Undefined();
-    }
-
-    Napi::Object out = Napi::Object::New(env);
-    Napi::Array layers = Napi::Array::New(env);
-    std::set<mapnik::vector_tile_impl::validity_error> errors;
-    bool has_errors = false;
-    std::size_t layers_size = 0;
-    bool first_layer = true;
-    std::set<std::string> layer_names_set;
-    std::uint32_t version = 1;
-    protozero::pbf_reader tile_msg;
-    std::string decompressed;
-    try
-    {
-        if (mapnik::vector_tile_impl::is_gzip_compressed(obj.As<Napi::Buffer<char>>().Data(),obj.As<Napi::Buffer<char>>().Length()) ||
-            mapnik::vector_tile_impl::is_zlib_compressed(obj.As<Napi::Buffer<char>>().Data(),obj.As<Napi::Buffer<char>>().Length()))
-        {
-            mapnik::vector_tile_impl::zlib_decompress(obj.As<Napi::Buffer<char>>().Data(), obj.As<Napi::Buffer<char>>().Length(), decompressed);
-            tile_msg = protozero::pbf_reader(decompressed);
-        }
-        else
-        {
-            tile_msg = protozero::pbf_reader(obj.As<Napi::Buffer<char>>().Data(),obj.As<Napi::Buffer<char>>().Length());
-        }
-        while (tile_msg.next())
-        {
-            switch (tile_msg.tag())
-            {
-                case mapnik::vector_tile_impl::Tile_Encoding::LAYERS:
-                    {
-                        Napi::Object layer_obj = Napi::Object::New(env);
-                        std::uint64_t point_feature_count = 0;
-                        std::uint64_t line_feature_count = 0;
-                        std::uint64_t polygon_feature_count = 0;
-                        std::uint64_t unknown_feature_count = 0;
-                        std::uint64_t raster_feature_count = 0;
-                        auto layer_view = tile_msg.get_view();
-                        protozero::pbf_reader layer_props_msg(layer_view);
-                        auto layer_info = mapnik::vector_tile_impl::get_layer_name_and_version(layer_props_msg);
-                        std::string const& layer_name = layer_info.first;
-                        std::uint32_t layer_version = layer_info.second;
-                        std::set<mapnik::vector_tile_impl::validity_error> layer_errors;
-                        if (version > 2 || version < 1)
-                        {
-                            layer_errors.insert(mapnik::vector_tile_impl::LAYER_HAS_UNSUPPORTED_VERSION);
-                        }
-                        protozero::pbf_reader layer_msg(layer_view);
-                        mapnik::vector_tile_impl::layer_is_valid(layer_msg,
-                                                                 layer_errors,
-                                                                 point_feature_count,
-                                                                 line_feature_count,
-                                                                 polygon_feature_count,
-                                                                 unknown_feature_count,
-                                                                 raster_feature_count);
-                        std::uint64_t feature_count = point_feature_count +
-                                                      line_feature_count +
-                                                      polygon_feature_count +
-                                                      unknown_feature_count +
-                                                      raster_feature_count;
-                        if (!layer_name.empty())
-                        {
-                            auto p = layer_names_set.insert(layer_name);
-                            if (!p.second)
-                            {
-                                errors.insert(mapnik::vector_tile_impl::TILE_REPEATED_LAYER_NAMES);
-                            }
-                            (layer_obj).Set(Napi::String::New(env, "name"), Napi::String::New(env, layer_name));
-                        }
-                        (layer_obj).Set(Napi::String::New(env, "features"), Napi::Number::New(env, feature_count));
-                        (layer_obj).Set(Napi::String::New(env, "point_features"), Napi::Number::New(env, point_feature_count));
-                        (layer_obj).Set(Napi::String::New(env, "linestring_features"), Napi::Number::New(env, line_feature_count));
-                        (layer_obj).Set(Napi::String::New(env, "polygon_features"), Napi::Number::New(env, polygon_feature_count));
-                        (layer_obj).Set(Napi::String::New(env, "unknown_features"), Napi::Number::New(env, unknown_feature_count));
-                        (layer_obj).Set(Napi::String::New(env, "raster_features"), Napi::Number::New(env, raster_feature_count));
-                        (layer_obj).Set(Napi::String::New(env, "version"), Napi::Number::New(env, layer_version));
-                        if (!layer_errors.empty())
-                        {
-                            has_errors = true;
-                            Napi::Array err_arr = Napi::Array::New(env);
-                            std::size_t i = 0;
-                            for (auto const& e : layer_errors)
-                            {
-                                (err_arr).Set(i++, Napi::String::New(env, mapnik::vector_tile_impl::validity_error_to_string(e)));
-                            }
-                            (layer_obj).Set(Napi::String::New(env, "errors"), err_arr);
-                        }
-                        if (first_layer)
-                        {
-                            version = layer_version;
-                        }
-                        else
-                        {
-                            if (version != layer_version)
-                            {
-                                errors.insert(mapnik::vector_tile_impl::TILE_HAS_DIFFERENT_VERSIONS);
-                            }
-                        }
-                        first_layer = false;
-                        (layers).Set(layers_size++, layer_obj);
-                    }
-                    break;
-                default:
-                    errors.insert(mapnik::vector_tile_impl::TILE_HAS_UNKNOWN_TAG);
-                    tile_msg.skip();
-                    break;
-            }
-        }
-    }
-    catch (...)
-    {
-        errors.insert(mapnik::vector_tile_impl::INVALID_PBF_BUFFER);
-    }
-    (out).Set(Napi::String::New(env, "layers"), layers);
-    has_errors = has_errors || !errors.empty();
-    (out).Set(Napi::String::New(env, "errors"),  Napi::Boolean::New(env, has_errors));
-    if (!errors.empty())
-    {
-        Napi::Array err_arr = Napi::Array::New(env);
-        std::size_t i = 0;
-        for (auto const& e : errors)
-        {
-            (err_arr).Set(i++, Napi::String::New(env, mapnik::vector_tile_impl::validity_error_to_string(e)));
-        }
-        (out).Set(Napi::String::New(env, "tile_errors"), err_arr);
-    }
-    return out;
-    return;
-    */
 }
