+++ src/mapnik_vector_tile.hpp
@@ -46,6 +46,7 @@ struct queryMany_result
 
 class VectorTile : public Napi::ObjectWrap<VectorTile>
 {
+    friend class Map;
 public:
     // initiaizer
     static Napi::Object Initialize(Napi::Env env, Napi::Object exports);
@@ -99,42 +100,6 @@ public:
     void set_tile_size(Napi::CallbackInfo const& info, const Napi::Value& value);
     Napi::Value get_buffer_size(Napi::CallbackInfo const& info);
     void set_buffer_size(Napi::CallbackInfo const& info, const Napi::Value& value);
-
-
-    void clear()
-    {
-        tile_->clear();
-    }
-
-    std::uint32_t tile_size() const
-    {
-        return tile_->tile_size();
-    }
-
-    void tile_size(std::uint32_t val)
-    {
-        tile_->tile_size(val);
-    }
-
-    std::int32_t buffer_size() const
-    {
-        return tile_->buffer_size();
-    }
-
-    void buffer_size(std::int32_t val)
-    {
-        tile_->buffer_size(val);
-    }
-
-    mapnik::vector_tile_impl::merc_tile_ptr & get_tile()
-    {
-        return tile_;
-    }
-
-    mapnik::vector_tile_impl::merc_tile_ptr const& get_tile() const
-    {
-        return tile_;
-    }
 
 private:
     static Napi::FunctionReference constructor;
